# openapi.yaml
openapi: 3.0.3
info:
  title: PixelProbe API
  # Version is dynamically set from version.py when served via /api/openapi.yaml
  version: 0.0.0
  description: |
    # PixelProbe REST API

    PixelProbe is a comprehensive media file corruption detection system that uses FFmpeg and other tools to scan and validate media files.

    Route and method coverage is checked in CI. Some response fields remain
    intentionally partial; consult docs/api.md and the response schemas before
    generating a strict client.

    ## Getting Started

    ### First-Time Setup

    1. Check if setup is required:
       ```bash
       curl -X GET http://your-server/api/auth/status
       ```

    2. If `first_run: true`, create admin account:
       ```bash
       curl -X POST http://your-server/api/auth/setup \
         -H "Content-Type: application/json" \
         -d '{"password": "your-secure-password"}'
       ```

    3. Login to get session cookie:
       ```bash
       curl -X POST http://your-server/api/auth/login \
         -H "Content-Type: application/json" \
         -d '{"username": "admin", "password": "your-secure-password"}' \
         -c cookies.txt
       ```

    ### Authentication Overview

    PixelProbe supports two authentication methods:

    #### Session-Based Authentication
    Used by the web interface. Login via `/auth/login` to receive a session cookie.

    #### API Token Authentication
    Recommended for programmatic access:

    1. Create an API token:
...

...

...

...

...

...

...

...

...

...
servers:
...
security:
  - bearerAuth: []
  - sessionAuth: []
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      description: Bearer token authentication using API tokens (opaque tokens generated by PixelProbe, not JWTs)
    sessionAuth:
      type: apiKey
      in: cookie
      name: session
      description: Session-based authentication via login
...
paths:
...
  /view/{result_id}:
    get:
      summary: View/stream file
      description: |
        Stream a media file for viewing. Supports range requests for video streaming.

        **Example:**
        ```bash
        curl -X GET http://your-server/api/view/123 \
          -H "Authorization: Bearer YOUR_TOKEN"
        ```
      tags: [Export]
      parameters:
        - name: result_id
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: File served successfully
        '206':
          description: Partial content (range request)
        '404':
          description: File not found
  /download/{result_id}:
    get:
      summary: Download file
      description: |
        Download a media file.

        **Example:**
        ```bash
        curl -X GET http://your-server/api/download/123 \
          -H "Authorization: Bearer YOUR_TOKEN" \
          --output file.mp4
        ```
      tags: [Export]
      parameters:
        - name: result_id
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: File download started
        '404':
          description: File not found
  /export:
    get:
      summary: Export scan results (GET)
      description: |
        Export scan results to CSV, JSON, or PDF via GET request.

        **Example:**
        ```bash
        curl -X GET "http://your-server/api/export?format=csv&filter=corrupted" \
          -H "Authorization: Bearer YOUR_TOKEN" \
          --output results.csv
        ```
      tags: [Export]
      parameters:
        - name: format
          in: query
          schema:
            type: string
            enum: [csv, json, pdf]
            default: csv
        - name: filter
          in: query
          schema:
            type: string
            enum: [all, corrupted, healthy, pending, error]
            default: all
        - name: search
          in: query
          schema:
            type: string
      responses:
        '200':
          description: Export file generated and returned

    post:
      summary: Export scan results (POST)
      description: |
        Export scan results to various formats (CSV, JSON, Excel, PDF).

        **Example:**
        ```bash
        curl -X POST http://your-server/api/export \
          -H "Authorization: Bearer YOUR_TOKEN" \
          -H "Content-Type: application/json" \
          -d '{
            "format": "csv",
            "filter": "corrupted"
          }' \
          --output results.csv
        ```
      tags: [Export]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [format]
              properties:
                format:
                  type: string
                  enum: [csv, json, pdf]
                filter:
                  type: string
                  enum: [all, corrupted, healthy, pending, warning]
                search:
                  type: string
                file_ids:
                  type: array
                  items:
                    type: integer
      responses:
        '200':
          description: Export file generated and returned

  # Maintenance Endpoints
...
  /scan-reports/{report_id}/export:
    get:
      summary: Export scan report as JSON
      description: |
        Export scan report as JSON with full scan results.

        **Example:**
        ```bash
        curl -X GET http://your-server/api/scan-reports/abc-123-def/export \
          -H "Authorization: Bearer YOUR_TOKEN" \
          --output report.json
        ```
      tags: [Reports]
      parameters:
        - name: report_id
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Report exported successfully
        '404':
          description: Report not found
...